+++ lib/tree_state_machine.dart
@@ -2,20 +2,20 @@
 library tree_state_machine;
 
 export 'src/machine/extensions.dart';
+export 'src/machine/tree_node.dart' show TreeNodeInfo;
 export 'src/machine/tree_state.dart'
     hide
-        TreeState,
-        DelegatingTreeState,
         DataTreeState,
         DelegatingDataTreeState,
-        NestedMachineState,
+        DelegatingTreeState,
         GoToResult,
         InternalTransitionResult,
+        NestedMachineState,
         SelfTransitionResult,
+        StateCreator,
         StopResult,
-        UnhandledResult,
-        StateCreator;
-export 'src/machine/initial_state_data.dart' hide InitialStateData;
+        TreeState,
+        UnhandledResult;
 export 'src/machine/data_value.dart' hide ClosableDataValue, VoidDataValue;
 export 'src/machine/tree_state_machine.dart'
     hide TestableTreeStateMachine, EncodableState, EncodableTree;